Move user page code into single file
